LayoutTests/http/tests/security/window-name-after-same-origin-aux-frame-navigation.html

<!DOCTYPE html>
<html>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
            testRunner.setCanOpenWindows(true);
        }

        console.log("window.name should equal \"test\" after a same-origin auxiliary frame navigation.");
        window.open("resources/window-name-test.html?127.0.0.1");
    </script>
</html>
