LayoutTests/http/tests/security/window-name-after-same-origin-aux-frame-navigation-expected.txt

CONSOLE MESSAGE: window.name should equal "test" after a same-origin auxiliary frame navigation.
CONSOLE MESSAGE: window.name equals "test".
